+++ tracker/maintenance/frmTrain.cs
@@ -37,70 +37,80 @@ namespace tracker.maintenance
         }
         private void btnAdd_Click(object sender, EventArgs e)
         {
-            if (btnAdd.Text == "&Add")
+            if (MysqlHelper.TestConnection())
             {
-                toggleInput(true);
-                btnAdd.Text = "&Save";
-                btnEdit.Enabled = false;
-                btnDelete.Enabled = false;
-                btnFind.Enabled = false;
-                btnCancel.Enabled = true;
-
-                txtCode.Text = "";
-                txtDesc.Text = "";
-                txtCode.Focus();
-
-                
-            }else
-            {
-                if (txtCode.Text == "")
+                if (btnAdd.Text == "&Add")
                 {
-                    MessageBox.Show("Code is required!", "Validation Error", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    toggleInput(true);
+                    btnAdd.Text = "&Save";
+                    btnEdit.Enabled = false;
+                    btnDelete.Enabled = false;
+                    btnFind.Enabled = false;
+                    btnCancel.Enabled = true;
+
+                    txtCode.Text = "";
+                    txtDesc.Text = "";
                     txtCode.Focus();
-                }else if (txtDesc.Text == "")
-                {
-                    MessageBox.Show("Description is required!", "Validation Error", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                    txtDesc.Focus();
+
+
                 }
                 else
                 {
-                    DataTable dt = MysqlHelper.ExecuteDataTable("SELECT * FROM trains where train_code ='" + txtCode.Text.Trim() + "'");
-                    if (dt.Rows.Count > 0)
+                    if (txtCode.Text == "")
                     {
-                        MessageBox.Show("Train code already exist.", this.Text, MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        MessageBox.Show("Code is required!", "Validation Error", MessageBoxButtons.OK, MessageBoxIcon.Information);
                         txtCode.Focus();
+                    }
+                    else if (txtDesc.Text == "")
+                    {
+                        MessageBox.Show("Description is required!", "Validation Error", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        txtDesc.Focus();
                     }
                     else
                     {
-                        if (addTrain() == 1)
+                        DataTable dt = MysqlHelper.ExecuteDataTable("SELECT * FROM trains where train_code ='" + txtCode.Text.Trim() + "'");
+                        if (dt.Rows.Count > 0)
                         {
-                            resetForm();
-
-                            dgvTrain.ClearSelection();//If you want
-
-                            int nRowIndex = dgvTrain.Rows.Count - 1;
-                            int nColumnIndex = 3;
-
-                            dgvTrain.Rows[nRowIndex].Selected = true;
-                            dgvTrain.Rows[nRowIndex].Cells[nColumnIndex].Selected = true;
-
-                            //In case if you want to scroll down as well.
-                            dgvTrain.FirstDisplayedScrollingRowIndex = nRowIndex;
+                            MessageBox.Show("Train code already exist.", this.Text, MessageBoxButtons.OK, MessageBoxIcon.Error);
+                            txtCode.Focus();
                         }
                         else
                         {
-                            MessageBox.Show("Unable to save transaction!", "Transaction Error", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                            txtCode.Focus();
+                            if (addTrain() == 1)
+                            {
+                                resetForm();
+
+                                dgvTrain.ClearSelection();//If you want
+
+                                int nRowIndex = dgvTrain.Rows.Count - 1;
+                                int nColumnIndex = 3;
+
+                                dgvTrain.Rows[nRowIndex].Selected = true;
+                                dgvTrain.Rows[nRowIndex].Cells[nColumnIndex].Selected = true;
+
+                                //In case if you want to scroll down as well.
+                                dgvTrain.FirstDisplayedScrollingRowIndex = nRowIndex;
+                            }
+                            else
+                            {
+                                MessageBox.Show("Unable to save transaction!", "Transaction Error", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                                txtCode.Focus();
+                            }
+                            //MessageBox.Show(addTrain().ToString());
+                            resetForm();
                         }
-                        //MessageBox.Show(addTrain().ToString());
-                        resetForm();
+
                     }
-                    
+                    //toggleInput(false);
+                    //btnAdd.Text = "&Add";
                 }
-                //toggleInput(false);
-                //btnAdd.Text = "&Add";
-            }
             
+            }
+            else
+            {
+                Application.Exit();
+            }
+           
         }
 
         private void btnCancel_Click(object sender, EventArgs e)
@@ -120,12 +130,17 @@ namespace tracker.maintenance
 
         private void loadGrid()
         {
-            dgvTrain.AutoGenerateColumns = false;
-            bs.DataSource = MysqlHelper.ExecuteDataTable("SELECT * FROM trains");
-            bdgNavigator.BindingSource = bs;
-            dgvTrain.DataSource = bs;
-
-
+            if (MysqlHelper.TestConnection())
+            {
+                dgvTrain.AutoGenerateColumns = false;
+                bs.DataSource = MysqlHelper.ExecuteDataTable("SELECT * FROM trains");
+                bdgNavigator.BindingSource = bs;
+                dgvTrain.DataSource = bs;
+            }
+            else
+            {
+                Application.Exit();
+            }
         }
 
         private void frmTrain_Load(object sender, EventArgs e)
@@ -166,38 +181,46 @@ namespace tracker.maintenance
 
         private void btnEdit_Click(object sender, EventArgs e)
         {
-            if (btnEdit.Text == "&Edit")
+            if (MysqlHelper.TestConnection())
             {
-                toggleInput(true);
-                btnEdit.Text = "&Update";
-                btnAdd.Enabled = false;
-                btnDelete.Enabled = false;
-                btnFind.Enabled = false;
-                btnCancel.Enabled = true;
-                txtCode.Focus();
-                int id = Int32.Parse(dgvTrain.CurrentRow.Cells["id"].Value.ToString());
-                DataTable dt = MysqlHelper.ExecuteDataTable("SELECT * FROM train_radios WHERE radio_id ='" + id + "'");
-                if (dt.Rows.Count > 0)
+                if (btnEdit.Text == "&Edit")
                 {
-                    chkActive.Enabled = false;
+                    toggleInput(true);
+                    btnEdit.Text = "&Update";
+                    btnAdd.Enabled = false;
+                    btnDelete.Enabled = false;
+                    btnFind.Enabled = false;
+                    btnCancel.Enabled = true;
+                    txtCode.Focus();
+                    int id = Int32.Parse(dgvTrain.CurrentRow.Cells["id"].Value.ToString());
+                    DataTable dt = MysqlHelper.ExecuteDataTable("SELECT * FROM train_radios WHERE radio_id ='" + id + "'");
+                    if (dt.Rows.Count > 0)
+                    {
+                        chkActive.Enabled = false;
+                    }
+                    else
+                    {
+                        chkActive.Enabled = true;
+                    }
                 }
                 else
                 {
-                    chkActive.Enabled = true;
+
+                    MysqlHelper.ExecuteNonQuery("UPDATE trains SET train_code=@code,train_desc=@desc, image_index=@image_index, active=@active WHERE id=@id",
+                    new MySqlParameter("@code", txtCode.Text.Trim()),
+                    new MySqlParameter("@desc", txtDesc.Text.Trim()),
+                    new MySqlParameter("@image_index", cmbIcon.SelectedIndex),
+                    new MySqlParameter("@active", chkActive.Checked ? 1 : 0),
+                    new MySqlParameter("@id", Convert.ToInt32(dgvTrain.CurrentRow.Cells[0].Value.ToString())));
+
+                    resetForm();
                 }
             }
             else
             {
-
-                MysqlHelper.ExecuteNonQuery("UPDATE trains SET train_code=@code,train_desc=@desc, image_index=@image_index, active=@active WHERE id=@id",
-                new MySqlParameter("@code", txtCode.Text.Trim()),
-                new MySqlParameter("@desc", txtDesc.Text.Trim()),
-                new MySqlParameter("@image_index", cmbIcon.SelectedIndex),
-                new MySqlParameter("@active", chkActive.Checked ? 1 : 0),
-                new MySqlParameter("@id", Convert.ToInt32(dgvTrain.CurrentRow.Cells[0].Value.ToString())));
-
-                resetForm();
+                Application.Exit();
             }
+            
         }
 
         private void btnClose_Click(object sender, EventArgs e)
@@ -207,57 +230,73 @@ namespace tracker.maintenance
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            DialogResult result = MessageBox.Show("Are you sure you want to delete this record?", this.Text,
-               MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-            if (result == DialogResult.Yes)
+            if (MysqlHelper.TestConnection())
             {
-                int id = Int32.Parse(dgvTrain.CurrentRow.Cells["id"].Value.ToString());
-                DataTable dt = MysqlHelper.ExecuteDataTable("SELECT * FROM train_radios WHERE train_id ='" + id + "'");
-                if (dt.Rows.Count == 0)
+                DialogResult result = MessageBox.Show("Are you sure you want to delete this record?", this.Text,
+                MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                if (result == DialogResult.Yes)
                 {
-                    int retVal = MysqlHelper.ExecuteNonQuery("DELETE FROM trains where id ='" + id + "'");
-                    if (retVal == 1)
+                    int id = Int32.Parse(dgvTrain.CurrentRow.Cells["id"].Value.ToString());
+                    DataTable dt = MysqlHelper.ExecuteDataTable("SELECT * FROM train_radios WHERE train_id ='" + id + "'");
+                    if (dt.Rows.Count == 0)
                     {
-                        loadGrid();
+                        int retVal = MysqlHelper.ExecuteNonQuery("DELETE FROM trains where id ='" + id + "'");
+                        if (retVal == 1)
+                        {
+                            loadGrid();
+                        }
+                        else
+                        {
+                            MessageBox.Show("Erorr deleting record.", this.Text, MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        }
                     }
                     else
                     {
-                        MessageBox.Show("Erorr deleting record.", this.Text, MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        MessageBox.Show("Cannot delete train it is currently assigned with a radio.", this.Text, MessageBoxButtons.OK, MessageBoxIcon.Error);
                     }
+
                 }
-                else
-                {
-                    MessageBox.Show("Cannot delete train it is currently assigned with a radio.", this.Text, MessageBoxButtons.OK, MessageBoxIcon.Error);
-                }
-                
             }
+            else
+            {
+                Application.Exit();
+            }
+            
         }
 
         private void btnFind_Click(object sender, EventArgs e)
         {
-            int result;
-            DataTable dt = MysqlHelper.ExecuteDataTable("SELECT id, train_code, train_desc FROM trains");
-
-            string[] fields = new string[] { "ID", "CODE", "DESCRIPTION" };
-            string[] field_ids = new string[] { "id", "train_code", "train_desc" };
-            int[] fieldsSize = new int[] { 50, 100, 400 };
-
-            frmFind frmFind = new frmFind();
-            frmFind.DataSource = dt;
-            frmFind.SearchFor = "Trains";
-            frmFind.FieldId = "id";
-            frmFind.Fields = fields;
-            frmFind.FieldIds = field_ids;
-            frmFind.FieldsSize = fieldsSize;
-
-            if (frmFind.ShowDialog() == DialogResult.OK)
+            if (MysqlHelper.TestConnection())
             {
-                result = frmFind.FilterValue;
-                loadGrid();
-                frmFind.MoveCursor(result, dgvTrain);
+                int result;
+                DataTable dt = MysqlHelper.ExecuteDataTable("SELECT id, train_code, train_desc FROM trains");
+
+                string[] fields = new string[] { "ID", "CODE", "DESCRIPTION" };
+                string[] field_ids = new string[] { "id", "train_code", "train_desc" };
+                int[] fieldsSize = new int[] { 50, 100, 400 };
+
+                frmFind frmFind = new frmFind();
+                frmFind.DataSource = dt;
+                frmFind.SearchFor = "Trains";
+                frmFind.FieldId = "id";
+                frmFind.Fields = fields;
+                frmFind.FieldIds = field_ids;
+                frmFind.FieldsSize = fieldsSize;
+
+                if (frmFind.ShowDialog() == DialogResult.OK)
+                {
+                    result = frmFind.FilterValue;
+                    loadGrid();
+                    frmFind.MoveCursor(result, dgvTrain);
+                }
+                frmFind.Dispose();
+                btnAdd.Focus();
             }
-            frmFind.Dispose();
-            btnAdd.Focus();
+            else
+            {
+                Application.Exit();
+            }
+            
         }
     }
 }
